+++ R/mexihat.R
@@ -48,7 +48,7 @@
 #' plot(mh$x, mh$psi, type="l", main = "Mexican Hat Wavelet",
 #'      xlab = "", ylab = "")
 #'
-#' @author Sylvain Pelissier, \email{<sylvain.pelissier@@gmail.com>}.\cr
+#' @author Sylvain Pelissier, \email{sylvain.pelissier@@gmail.com}.\cr
 #' Conversion to R by Geert van Boxtel, \email{G.J.M.vanBoxtel@@gmail.com}.
 #
 #' @export