+++ tests/testthat/test_Sample_Rate_Change_Functions.R
@@ -15,6 +15,10 @@ test_that("parameters to upfirdn() are correct", {
 test_that("upfirdn() tests are correct", {
   expect_equal(upfirdn(1:100, 1, 1, 1), seq(1, 100, 1))
   expect_equal(upfirdn(1:100, 1, 1, 2), seq(1, 100, 2))
+  expect_equal(upfirdn(cbind(1:100, 1:100), 1, 1, 1),
+               cbind(seq(1, 100, 1), seq(1, 100, 1)))
+  expect_equal(upfirdn(cbind(1:100, 1:100), 1, 1, 2),
+               cbind(seq(1, 100, 2), seq(1, 100, 2)))
 })
 
 # -----------------------------------------------------------------------
@@ -31,8 +35,12 @@ test_that("parameters to resample() are correct", {
 
 test_that("resample() tests are correct", {
   expect_equal(resample(1:100, 1, 1), seq(1, 100, 1))
+  expect_equal(resample(cbind(1:100, 1:100), 1, 1),
+               cbind(seq(1, 100, 1), seq(1, 100, 1)))
   expect_equal(length(resample(1:100, 1, 2)), 50)
+  expect_equal(nrow(resample(cbind(1:100, 1:100), 1, 2)), 50)
   expect_equal(length(resample(1:100, 2, 1)), 200)
+  expect_equal(nrow(resample(cbind(1:100, 1:100), 2, 1)), 200)
 })
 
 # -----------------------------------------------------------------------
@@ -75,4 +83,25 @@ test_that("upsample() tests are correct", {
   expect_equal(upsample(c(2, 4), 2, 1), c(0, 2, 0, 4))
   expect_equal(upsample(matrix(c(3, 4, 7, 8), 2, byrow = TRUE), 2, 1),
                matrix(c(0, 0, 3, 4, 0, 0, 7, 8), 4, byrow = TRUE))
+})
+
+# -----------------------------------------------------------------------
+# decimate()
+
+test_that("parameters to upsample() are correct", {
+  expect_error(decimate())
+  expect_error(decimate(1))
+  expect_error(decimate(1, -1))
+  expect_error(decimate(1, 1, -1))
+  expect_error(decimate(1, 2, 3, 4))
+  expect_error(decimate(1, 2, 3, "error"))
+})
+
+test_that("upsample() tests are correct", {
+  expect_equal(round(decimate(1:10, 2), 3), c(1.306, 2.833, 5.037, 6.789, 9.142))
+  expect_equal(round(decimate(1:10, 2, ftype = "fir"), 3),
+               c(-0.002, -0.002, -0.007, -0.003, -0.014))
+  expect_equal(round(decimate(matrix(c(1:10, 1:10), ncol = 2), 2), 3),
+               matrix(c(1.306, 2.833, 5.037, 6.789, 9.142,
+                        1.306, 2.833, 5.037, 6.789, 9.142), ncol = 2))
 })